+++ packages/terriajs/test/Map/StyledHtmlSpec.tsx
@@ -1,11 +1,10 @@
 import { render, screen } from "@testing-library/react";
 import { runInAction } from "mobx";
-import { ThemeProvider } from "styled-components";
 import Terria from "../../lib/Models/Terria";
 import ViewState from "../../lib/ReactViewModels/ViewState";
-import { terriaTheme } from "../../lib/ReactViews/StandardUserInterface";
 import { StyledHtmlRaw } from "../../lib/ReactViews/Map/Panels/HelpPanel/StyledHtml";
 import registerCustomComponentTypes from "../../lib/ReactViews/Custom/registerCustomComponentTypes";
+import { TerriaThemeProvider } from "../ReactViews/withContext";
 
 describe("StyledHtml", function () {
   let terria: Terria;
@@ -23,12 +22,12 @@ describe("StyledHtml", function () {
   describe("with basic props", function () {
     it("mounts without problems", function () {
       render(
-        <ThemeProvider theme={terriaTheme}>
+        <TerriaThemeProvider>
           <StyledHtmlRaw
             markdown={"something something spatial data mochi"}
             viewState={viewState}
           />
-        </ThemeProvider>
+        </TerriaThemeProvider>
       );
 
       expect(
@@ -50,12 +49,12 @@ describe("StyledHtml", function () {
         });
       });
       render(
-        <ThemeProvider theme={terriaTheme}>
+        <TerriaThemeProvider>
           <StyledHtmlRaw
             markdown={"something something spatial data mochi"}
             viewState={viewState}
           />
-        </ThemeProvider>
+        </TerriaThemeProvider>
       );
 
       expect(